% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diceRoller.R
\name{dice_roller}
\alias{dice_roller}
\title{Dice Roller
mimics rolling a die or dice of the specified size (d6, d12, etc.).}
\usage{
dice_roller(faces, dice = 1, add = TRUE)
}
\arguments{
\item{faces}{The number of sides or faces of the die you wish to roll.}

\item{dice}{The number of dice you wish to roll. The default is 1.}

\item{add}{For rolls with more than one dice, whether you wish to add the dice together (TRUE) or get the mean, maximum, and minimum of the dice (FALSE). The default is TRUE.}
}
\value{
Either one number (the result of the die rolled or the result of the dice rolled then added together) or three numbers (the mean, maximum, and minimum of the dice rolled).
}
\description{
Dice Roller
mimics rolling a die or dice of the specified size (d6, d12, etc.).
}
\examples{
dice_roller(20)
dice_roller(6, 4)
dice_roller(4, 3, FALSE)
}
