% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll polyhedral dice}
\usage{
roll(dice = 1, sides = 20, bonus = 0)
}
\arguments{
\item{dice}{The number of dice to roll. Defaults to 1}

\item{sides}{The number of sides on each die. Defaults to 20.}

\item{bonus}{The fixed bonus to add to the result. Defaults to 0.}
}
\value{
Strings reflecting the number of dice rolled of which kinds, the number rolled, and the total with bonus.
}
\description{
Roll X dice with Y sides and add a fixed bonus.
}
\examples{
roll()

roll(2, 6)

roll(3, 8, 2)
}
