% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolladv.R
\name{rolladv}
\alias{rolladv}
\title{Roll with advantage}
\usage{
rolladv(advantage = T, bonus = 0)
}
\arguments{
\item{advantage}{Boolean. If FALSE will roll with disadvantage (lower of two dice), otherwise will roll with advantage (better of two dice).}

\item{bonus}{A fixed value to add to the final result.}
}
\value{

}
\description{
Rolls a d20 with the advantage or disadvantage mechanic from Dungeons & Dragons 5th Edition, where you roll two d20s and take the better or worse result.
}
\examples{
rolladv()

rolladv(F)

rolladv(,2)

rolladv(F, 4)
}
