% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolld.R
\name{rolld}
\alias{rolld}
\title{Roll XdY+Z dice}
\usage{
rolld(diceroll)
}
\arguments{
\item{diceroll}{A number of dice to roll, in a string with the format "XdY" or "XdY+Z", where X = the number of dice, Y = the number of sides per dice, and Z = any bonus to the roll.}
}
\value{
String restating the dice roll, giving the numbers rolled, and showing the total value with bonuses
}
\description{
When given an input in the form of a string "XdY" or "XdY+Z", where X = the number of dice, Y = the number of sides per dice, and Z = any bonus to the roll, computes the result of that die roll.
}
\examples{
diceroll("3d6+2")
}
